fix(FileSystem): toUri for default fs
`JnmofsFileSystem.toUri` would fail with URI syntax exception when used
on the default jnmofs filesystem ("") as the filesystem uri.getHost()
returns null in this case. Guard against that and use an empty string
hostname in that case instead.

Also fixed the test to correctly expect jnmofs://// as URI for / on the
default FileSystem, which is consistent with other expectations.