fix(JnmofsRegularFile): manage buffer limit when dumping contents
Source buffer limit wasn't properly managed which could lead to a
BufferUnderflowException.

Also, had to test ZipFs provider using its instance directly instead of
going through `FileSystems.newFileSystem` as the latter swallows
`UnsupportedOperationException` which would lead to the Zip file
creation failing with `provider "jar" not found` errors when it was
actually due to some method in the underlying filesystem being not
implemented yet. All methods required so far have already been
implemented, but leaving this in for documentation, or in case another
such method is encountered later.